+++ src/main/java/com/kirillrublevsky/dao/impl/DollarsDAOImpl.java
@@ -1,7 +1,7 @@
 package com.kirillrublevsky.dao.impl;
 
 import com.kirillrublevsky.dao.DollarsDAO;
-import com.kirillrublevsky.model.Dollars;
+import com.kirillrublevsky.dao.entity.Dollars;
 import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
@@ -15,12 +15,12 @@ public class DollarsDAOImpl implements DollarsDAO {
     public DollarsDAOImpl() {}
 
     @Override
-    public void updateDollars(Integer amount) {
+    public void updateDollars(double amount) {
         Dollars dollars = (Dollars) sessionFactory.getCurrentSession().load(Dollars.class, 1);
         dollars.setAmount(amount);    }
 
     @Override
-    public Integer getDollars(){
+    public double getDollars(){
         Dollars dollars = (Dollars) sessionFactory.getCurrentSession().load(Dollars.class, 1);
         return dollars.getAmount();
     }
