+++ src/main/java/com/kirillrublevsky/service/DollarsService.java
@@ -2,6 +2,6 @@ package com.kirillrublevsky.service;
 
 public interface DollarsService {
 
-    void updateDollars(Integer amount);
-    Integer getDollars();
+    void updateDollars(double amount);
+    double getDollars();
 }
